add skip wasm test
