+++ pkg/app/robot.txt
@@ -1,0 +1,1 @@
+hello
